Moved contents of Search Controller to Recipe Controller
